pyproject.toml
[tool.poetry]
name = "marketplace"
version = "0.1.0"
description = ""
authors = ["hmbefi <hmbefi@email.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
pydantic-settings = "^2.1.0"
sqlalchemy = "^2.0.23"
alembic = "^1.12.1"
asyncpg = "^0.29.0"
python-multipart = "^0.0.6"

[tool.poetry.group.lint]
optional = true

[tool.poetry.group.lint.dependencies]
wemake-python-styleguide = "^0.18.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
